Add files via upload
This folder contains a personal project in which I used a coding tool called Cubic to reverse-engineer and customize a generic Ubuntu 20.04.6 LTS ISO.  I used Cubic to code a script, created a C program to run the script, and created a desktop application program to automatically run the C program on boot.  Check out the "how_to_create_cubic_iso" to learn how to do it!